Moved treeview into separate control
